Conflicts:
	Music_Thing/nbproject/project.properties

+++ Music_Thing/src/music_thing/FXMLDocumentController.java
@@ -16,7 +16,10 @@ import java.util.List;
 import java.util.Optional;
 import java.util.ResourceBundle;
 import javafx.application.Platform;
+import javafx.beans.value.ChangeListener;
+import javafx.beans.value.ObservableValue;
 import javafx.event.ActionEvent;
+import javafx.event.Event;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.Alert;
@@ -27,6 +30,7 @@ import javafx.scene.control.Label;
 import javafx.scene.control.MenuItem;
 import javafx.scene.control.SelectionMode;
 import javafx.scene.control.Slider;
+import javafx.scene.control.TableCell;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.cell.PropertyValueFactory;
@@ -36,6 +40,7 @@ import javafx.scene.input.MouseEvent;
 import javafx.scene.input.TransferMode;
 import javafx.scene.layout.HBox;
 import javafx.scene.shape.Polygon;
+import javafx.util.Callback;
 import javax.swing.JFileChooser;
 import javax.swing.SwingUtilities;
 import javax.swing.filechooser.FileNameExtensionFilter;
@@ -78,6 +83,10 @@ public class FXMLDocumentController implements Initializable {
     private HBox pauseSymbol;
     @FXML
     private Polygon playSymbol;
+    @FXML
+    private MenuItem menuPlay;
+    
+    private Main main;
     
     private MusicPlayer player;
     private final JavafxPlayer jfxPlayer = new JavafxPlayer();
@@ -92,18 +101,43 @@ public class FXMLDocumentController implements Initializable {
     }
     
     @FXML
-    private void stopMusic(MouseEvent event) {
+    private void stopMusic(Event event) {
         if(player!=null)player.reset();
         pauseSymbol.setVisible(false);
         playSymbol.setVisible(true);
+        menuPlay.setText("Play");
         songList.requestFocus();
     }
     
     @FXML
-    private void play(MouseEvent event) {
+    private void nextSong(Event event){
+        if(MusicLibrary.getTrackNumber()<MusicLibrary.size()-1){
+            songList.getSelectionModel().clearAndSelect(MusicLibrary.getTrackNumber()+1);
+            MusicLibrary.setTrack(MusicLibrary.getTrackNumber()+1);
+            if(player!=null && player.getPlaying()==true)play(event);
+        }else{
+            stopMusic(event);
+        }
+        songList.requestFocus();
+    }
+    
+    @FXML
+    private void prevSong(Event event){
+        if(MusicLibrary.getTrackNumber()>0){
+            songList.getSelectionModel().clearAndSelect(MusicLibrary.getTrackNumber()-1);
+            MusicLibrary.setTrack(MusicLibrary.getTrackNumber()-1);
+            if(player!=null && player.getPlaying()==true)play(event);
+        }else{
+            stopMusic(event);
+        }
+        songList.requestFocus();
+    }
+    
+    @FXML
+    private void play(Event event) {
         if(MusicLibrary.size()>0){
             Track selectedTrack = MusicLibrary.getSelectedTrack(songList);
-            if(selectedTrack!=player.getCurrentTrack()){
+            if(player!=null && selectedTrack!=player.getCurrentTrack() || player==null){
                 SongType type = selectedTrack.getType();
                 if(player!=null)player.stop();
                 if(type==SongType.MP3 || type==SongType.AAC || type==SongType.WAV || type==SongType.AIFF){
@@ -118,14 +152,17 @@ public class FXMLDocumentController implements Initializable {
                 player.play(selectedTrack, songVolumeBar.getValue());
                 pauseSymbol.setVisible(true);
                 playSymbol.setVisible(false);
+                menuPlay.setText("Pause");
             }else if(player.getPlaying() && MusicLibrary.getSelectedTrack(songList)!=player.getCurrentTrack()){
                 player.play(selectedTrack, songVolumeBar.getValue());
                 pauseSymbol.setVisible(true);
                 playSymbol.setVisible(false);
+                menuPlay.setText("Pause");
             }else{
                 player.pause();
                 pauseSymbol.setVisible(false);
                 playSymbol.setVisible(true);
+                menuPlay.setText("Play");
             }
             defaultTimeLabel.setVisible(false);
             songTime.setVisible(true);
@@ -142,12 +179,12 @@ public class FXMLDocumentController implements Initializable {
     }
     
     @FXML
-    private void changeVolume(MouseEvent event){
+    private void changeVolume(Event event){
         if(player!=null)player.setVolume(songVolumeBar.getValue());
     }
     
     @FXML
-    private void deleteFile(ActionEvent event){
+    private void deleteFile(Event event){
         Platform.runLater(() -> {
             try{
                 HashSet<Track> toDelete = new HashSet(songList.getSelectionModel().getSelectedItems());
@@ -159,19 +196,17 @@ public class FXMLDocumentController implements Initializable {
                     if(toDelete.size()==1)alert.setContentText("Are you sure you want to delete 1 track?");
                     Optional<ButtonType> result = alert.showAndWait();
                     if (result.get() == ButtonType.OK){
-                        if(toDelete.contains(player.getCurrentTrack())){
+                        if(player!=null && toDelete.contains(player.getCurrentTrack())){
                             player.stop();
                         }
                         for(Track track: toDelete){
-                            //Files.delete(Paths.get("music/"+toDelete.getPath()));
-                            //MusicLibrary.removeTrack(toDelete);
                             Files.delete(Paths.get("music/"+track.getPath()));
                             MusicLibrary.removeTrack(track);
                         }
                         MusicLibrary.setTrack(songList.getFocusModel().getFocusedCell().getRow());
                     }
                 }
-            }catch(Exception e){System.out.println(e);}
+            }catch(Exception e){}
             MusicLibrary.save();
         });
     }
@@ -245,26 +280,24 @@ public class FXMLDocumentController implements Initializable {
         if(file.isFile()){
             File copyTo =  new File("music/"+file.getName());
             try{
-                if(!copyTo.exists()){
-                    if(file.getName().toLowerCase().endsWith("mp3")){
-                        Files.copy(file.toPath(), copyTo.toPath());
-                        MusicLibrary.addSong(new Track(SongType.MP3, file.getName()));
-                    }else if(file.getName().toLowerCase().endsWith("mid")){
-                        Files.copy(file.toPath(), copyTo.toPath());
-                        MusicLibrary.addSong(new Track(SongType.MIDI, file.getName()));
-                    }else if(file.getName().toLowerCase().endsWith("m4a")){
-                        Files.copy(file.toPath(), copyTo.toPath());
-                        MusicLibrary.addSong(new Track(SongType.AAC, file.getName()));
-                    }else if(file.getName().toLowerCase().endsWith("aiff")){
-                        Files.copy(file.toPath(), copyTo.toPath());
-                        MusicLibrary.addSong(new Track(SongType.AIFF, file.getName()));
-                    }else if(file.getName().toLowerCase().endsWith("wav")){
-                        Files.copy(file.toPath(), copyTo.toPath());
-                        MusicLibrary.addSong(new Track(SongType.WAV, file.getName()));
-                    }else if(file.getName().toLowerCase().endsWith("flac")){
-                        Files.copy(file.toPath(), copyTo.toPath());
-                        MusicLibrary.addSong(new Track(SongType.FLAC, file.getName()));
-                    }
+                if(file.getName().toLowerCase().endsWith("mp3")){
+                    if(!copyTo.exists())Files.copy(file.toPath(), copyTo.toPath());
+                    MusicLibrary.addSong(new Track(SongType.MP3, file.getName()));
+                }else if(file.getName().toLowerCase().endsWith("mid")){
+                    if(!copyTo.exists())Files.copy(file.toPath(), copyTo.toPath());
+                    MusicLibrary.addSong(new Track(SongType.MIDI, file.getName()));
+                }else if(file.getName().toLowerCase().endsWith("m4a")){
+                    if(!copyTo.exists())Files.copy(file.toPath(), copyTo.toPath());
+                    MusicLibrary.addSong(new Track(SongType.AAC, file.getName()));
+                }else if(file.getName().toLowerCase().endsWith("aiff")){
+                    if(!copyTo.exists())Files.copy(file.toPath(), copyTo.toPath());
+                    MusicLibrary.addSong(new Track(SongType.AIFF, file.getName()));
+                }else if(file.getName().toLowerCase().endsWith("wav")){
+                    if(!copyTo.exists())Files.copy(file.toPath(), copyTo.toPath());
+                    MusicLibrary.addSong(new Track(SongType.WAV, file.getName()));
+                }else if(file.getName().toLowerCase().endsWith("flac")){
+                    if(!copyTo.exists())Files.copy(file.toPath(), copyTo.toPath());
+                    MusicLibrary.addSong(new Track(SongType.FLAC, file.getName()));
                 }
             }catch (Exception e){}
             
@@ -272,6 +305,58 @@ public class FXMLDocumentController implements Initializable {
             for(File thing : file.listFiles()) importFile(thing);
         }
         
+    }
+    
+    /**
+     * Called when the user clicks the edit button.
+     * Opens a dialog to edit details for the selected track.
+     */
+    @FXML
+    private void handleEditTrack() {
+        if(MusicLibrary.size()>0){
+            Track track = MusicLibrary.getSelectedTrack(songList);
+            if (track != null) {
+              boolean okClicked = main.showTrackEditDialog(track);
+              if (okClicked) {
+                refresh();
+              }
+
+            } else {
+              // Nothing selected.
+              Alert alert = new Alert(AlertType.WARNING);
+              alert.initOwner(Main.getMainWindow());
+              alert.setTitle("No Selection");
+              alert.setHeaderText("No Track Selected");
+              alert.setContentText("Please select a track in the table.");
+
+              alert.showAndWait();
+            }
+        }
+    }
+
+    /**
+     * Refreshes the table. This is only necessary if an item that is already in
+     * the table is changed. New and deleted items are refreshed automatically.
+     * 
+     * This is a workaround because otherwise we would need to use property
+     * bindings in the model class and add a *property() method for each
+     * property. Maybe this will not be necessary in future versions of JavaFX
+     */
+    private void refresh() {
+      int selectedIndex = songList.getSelectionModel().getSelectedIndex();
+      songList.setItems(null);
+      songList.layout();
+      songList.setItems(MusicLibrary.getLibrary());
+      songList.getSelectionModel().select(selectedIndex);
+    }
+    
+    public void setMain(Main main){
+        this.main = main;
+    }
+    
+    @FXML
+    public void toggleFulllscreen(ActionEvent event){
+        Main.getMainWindow().setFullScreen(!Main.getMainWindow().isFullScreen());
     }
     
     @Override
@@ -293,6 +378,28 @@ public class FXMLDocumentController implements Initializable {
                 new PropertyValueFactory("playCount"));
         timeCol.setCellValueFactory(
                 new PropertyValueFactory("length"));
+                ratingCol.setCellFactory(new Callback<TableColumn<Track,Double>,TableCell<Track,Double>>() {
+                    @Override
+                    public TableCell<Track, Double> call(TableColumn<Track,Double> param){
+                        TableCell<Track, Double> cell = new TableCell<Track, Double>(){
+                            @Override
+                            public void updateItem(Double d, boolean empty){
+                                if(d != null){
+                                    Rating rating = new Rating();
+                                    rating.setRating(d.intValue());
+                                    rating.getRatingProperty().addListener(new ChangeListener(){
+                                        @Override
+                                        public void changed(ObservableValue observable, Object oldValue, Object newValue) {
+                                             MusicLibrary.getLibrary().get(getIndex()).setRating(((Integer)newValue).doubleValue());
+                                        }
+                                    });
+                                    setGraphic(rating);
+                                }
+                            }
+                        };
+                        return cell;
+                    }
+                }); 
         songList.setItems(MusicLibrary.getLibrary());
         songList.getSelectionModel().setSelectionMode(SelectionMode.MULTIPLE);
     }
